requirements.txt
bertopic
sentence-transformers
umap-learn
hdbscan
scikit-learn
seaborn
numpy==1.26.4
matplotlib==3.8.4
spacy==3.7.4
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
nltk==3.8.1
transformers==4.44.2
torch==2.3.1
accelerate==0.31.0
pyarrow==16.1.0
fastparquet==2024.5.0
datasets==2.20.0
peft==0.11.1
bitsandbytes==0.45.1
git+https://github.com/antonio258/npp.git